[Setting] Model 폴더 추가하여 Model 코드 옮겼습니다 (#32)
