+++ GetARock/GetARock/Global/UIComponent/PositionCollectionView.swift
@@ -7,13 +7,12 @@
 
 import UIKit
 
-enum Section: Int {
-    case main
+protocol PositionCollectionViewDelegate: AnyObject {
+    func canSelectPosition(_ collectionView: UICollectionView, indexPath: IndexPath, selectedItemsCount: Int) -> Bool
 }
 
-enum entryPoint {
-    case band
-    case exceptBand
+enum Section: Int {
+    case main
 }
 
 enum Item: Hashable {
@@ -21,39 +20,20 @@ enum Item: Hashable {
     case position(Position)
 }
 
-struct BandMember: Hashable {
-    let id = UUID()
-    let isUser: Bool
-    let isLeader: Bool
-    let userName: String
-    let imageName: String
-    let instrumentNames: [String]
-}
-
-struct Position: Hashable {
-    let id = UUID()
-    let instrumentName: String
-    let imageName: String
-    let isETC: Bool
-}
-//
-//포지션 선택
-//name: "기타",
-//imageName: "guitar",
-//isMain: true, // 가장 처음 선택한 것을 true
-//isETC: true // name중 기본 name에 없는 것을 etc로 판단할지, isetc를 넣을지? etc는 x를 넣어줘야함
-//
-
-// selection이 있는 경우 : 모두 선택 / 최대2개 -> protocol로 조절
-// selection이 없는 경우 : delegate를 쓰지 않음,
-// 추가버튼 있는 것과, + 버튼은 어떻게?
-protocol PositionCollectionViewDelegate: AnyObject {
-    func canSelectPosition(_ collectionView: UICollectionView, indexPath: IndexPath, selectedItemsCount: Int) -> Bool
-}
-
 final class PositionCollectionView: UIView {
     
-    // MARK: - properties
+    // MARK: - Property
+    
+    enum EntryPoint {
+        case band
+        case exceptBand
+    }
+    
+    private var entryPoint: EntryPoint
+    weak var delegate: PositionCollectionViewDelegate?
+    private lazy var dataSource: UICollectionViewDiffableDataSource<Section, Item> = self.makeDataSource()
+    
+    // MARK: - View
     
     private lazy var collectionView: UICollectionView = {
         let itemSize = NSCollectionLayoutSize(widthDimension: .absolute(174),
@@ -79,7 +59,7 @@ final class PositionCollectionView: UIView {
         collectionView.backgroundColor = .clear
         collectionView.allowsMultipleSelection = true
         collectionView.delegate = self
-        switch entryPoint {
+        switch self.entryPoint {
         case .band:
             collectionView.register(BandMemeberCollectionViewCell.self, forCellWithReuseIdentifier: BandMemeberCollectionViewCell.className)
         case .exceptBand:
@@ -87,10 +67,7 @@ final class PositionCollectionView: UIView {
         }
         return collectionView
     }()
-    private var entryPoint: entryPoint
-    private lazy var dataSource: UICollectionViewDiffableDataSource<Section, Item> = self.makeDataSource()
-    weak var delegate: PositionCollectionViewDelegate?
-    
+
     private var positions: [Item] = [.position(Position(instrumentName: "기타", imageName: "guitar", isETC: false)),
                                      .position(Position(instrumentName: "베이스", imageName: "bass", isETC: false)),
                                      .position(Position(instrumentName: "보컬", imageName: "vocal", isETC: false)),
@@ -99,7 +76,7 @@ final class PositionCollectionView: UIView {
     
     // MARK: - init
     
-    init(entryPoint: entryPoint) {
+    init(entryPoint: EntryPoint) {
         self.entryPoint = entryPoint
         super.init(frame: .zero)
         setupLayout()
@@ -109,7 +86,6 @@ final class PositionCollectionView: UIView {
         case .exceptBand:
             configurePositionView()
         }
-        
     }
     
     required init?(coder: NSCoder) {
@@ -127,7 +103,6 @@ final class PositionCollectionView: UIView {
 // MARK: - diffable
 
 extension PositionCollectionView {
-    
     private func makeDataSource() -> UICollectionViewDiffableDataSource<Section, Item> {
         return UICollectionViewDiffableDataSource<Section, Item>(collectionView: self.collectionView, cellProvider: { collectionView, indexPath, item in
             switch item {
@@ -140,22 +115,6 @@ extension PositionCollectionView {
                 cell.configure(data: position)
                 return cell
             }
-        })
-    }
-    
-    private func configureBandMemberDataSource() -> UICollectionViewDiffableDataSource<Section, BandMember> {
-        return UICollectionViewDiffableDataSource<Section, BandMember>(collectionView: self.collectionView, cellProvider: { collectionView, indexPath, bandMember in
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: BandMemeberCollectionViewCell.className, for: indexPath) as? BandMemeberCollectionViewCell else { return UICollectionViewCell() }
-            cell.configure(data: bandMember)
-            return cell
-        })
-    }
-    
-    private func configurePositionDataSource() -> UICollectionViewDiffableDataSource<Section, Position>{
-        return UICollectionViewDiffableDataSource<Section, Position>(collectionView: self.collectionView, cellProvider: { collectionView, indexPath, position in
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: PositionCollectionViewCell.className, for: indexPath) as? PositionCollectionViewCell else { return UICollectionViewCell() }
-            cell.configure(data: position)
-            return cell
         })
     }
     
@@ -179,7 +138,6 @@ extension PositionCollectionView {
 }
 
 extension PositionCollectionView: UICollectionViewDelegate {
-
     func collectionView(_ collectionView: UICollectionView, shouldSelectItemAt indexPath: IndexPath) -> Bool {
         let selectedPositionCount = collectionView.indexPathsForSelectedItems?.count ?? 0
         guard let canSelect = delegate?.canSelectPosition(collectionView, indexPath: indexPath, selectedItemsCount: selectedPositionCount) else { return false }
