+++ GetARock/GetARock/Network/Model/BandMember.swift
@@ -1,0 +1,17 @@
+//
+//  BandMember.swift
+//  GetARock
+//
+//  Created by 최동권 on 2023/01/24.
+//
+
+import Foundation
+
+struct BandMember: Hashable {
+    let id = UUID()
+    let isUser: Bool
+    let isLeader: Bool
+    let userName: String
+    let imageName: String
+    let instrumentNames: [String]
+}
