+++ GetARock/GetARock/Network/Model/Position.swift
@@ -1,0 +1,15 @@
+//
+//  Position.swift
+//  GetARock
+//
+//  Created by 최동권 on 2023/01/24.
+//
+
+import Foundation
+
+struct Position: Hashable {
+    let id = UUID()
+    let instrumentName: String
+    let imageName: String
+    let isETC: Bool
+}
